add POC calculation of h.264

+++ ClosedCaption/wv_h264pps.h
@@ -8,15 +8,15 @@
 
 typedef struct wv_h264pps
 {
-	uint8_t  m_pps_id;
-	uint8_t  m_sps_id;
+	int32_t  m_pps_id;
+	int32_t  m_sps_id;
 	bool     m_entropy_coding_flag;
 	bool     m_bottom_field_pic_order_in_frame_present_flag;
-	uint8_t  m_num_slice_groups;
-	uint8_t  m_num_ref_idx_l0_default_active;
-	uint8_t  m_num_ref_idx_l1_default_active;
+	int32_t  m_num_slice_groups;
+	int32_t  m_num_ref_idx_l0_default_active;
+	int32_t  m_num_ref_idx_l1_default_active;
 	bool     m_weighted_pred_flag;
-	uint8_t  m_weighted_bipred_idc;
+	int32_t  m_weighted_bipred_idc;
 	int32_t  m_pic_init_qp;
 	int32_t  m_pic_init_qs;
 	int32_t  m_chroma_qp_index_offset;
